wd-security-cli.c: rename switches that enable wdpassport-utils.py quirks
Rename the switches which enable and disable the wdpassport-utils.py
workaround quirks.  By default, the automatic detection of a Security
Block created by wdpassport-utils.py is enabled.  Instead of having
 --no-detect to disable the automatic detection, and --force-wdp-utils
to force enable the quirks, let's have --no-wdp-utils to disable it and
simply --wdp-utils to force enable it, which provides more symmetry in
my opinion.  The automatic detection is still enabled by default.