configure.ac: only try to include linux/fs.h if it exists
Currently the configure statement checking for BLKRRPART will try to
include linux/fs.h, but this header won't exists on non-Linux platforms.
Well, I guess BLKRRPART probably won't exist either, but this check
should fail because of the check for BLKRRPART not because of a failed
include statement.  Let's wrap the include statement with ifdef
HAVE_LINUX_FS_H.